% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit Conversion}
\usage{
C_to_F(temp_C)
}
\arguments{
\item{temp_C}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert temperature in degrees Celsius to temperature in degrees Fahrenheit
}
\examples{
temp1 <- C_to_F(100);
temp2 <- C_to_F(c(-5, 0, 100));
}
