% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius Conversion}
\usage{
F_to_C(temp_F)
}
\arguments{
\item{temp_F}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert temperature in degrees Fahrenheit to temperature in degrees Celsius
}
\examples{
temp1 <- F_to_C(32);
temp2 <- F_to_C(c(10, 32, 212));
}
